refactor: move background overlay into component
